Split load-examples.groovy into LoadPackages.groovy and LoadPackage.groovy
