+++ config/scheduler/settings.ini
@@ -83,7 +83,7 @@ pa.scheduler.core.starttask.threadnumber=5
 pa.scheduler.core.listener.threadnumber=5
 
 # List of the scripts paths to execute at scheduler start. Paths are separated by a ';'.
-pa.scheduler.startscripts.paths=tools/load-examples.groovy
+pa.scheduler.startscripts.paths=tools/LoadPackages.groovy
 
 #-------------------------------------------------------
 #----------------   JOBS PROPERTIES   ------------------
